refactor: 把番号解析并进 file_info, 测试收成同一文件里的表
